web: Extract frameless code highlighter
